osa4:4.23

+++ osa4/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -9,21 +7,13 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
   })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
   
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'user not authorized' })
   }
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog ({
     title: body.title,
@@ -32,6 +22,7 @@ blogsRouter.post('/', async (request, response) => {
     likes: body.likes,
     user : user._id
   })
+
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -48,6 +39,16 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!user) {
+    return response.status(401).json({ error: 'user not authorized' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'forbidden' }) 
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
